% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildfreqtable.R
\name{buildfreqtable}
\alias{buildfreqtable}
\title{Convert raw data to count data for use in snell function}
\usage{
buildfreqtable(data, trait, subgroup)
}
\arguments{
\item{data}{A data frame containing the raw data}

\item{trait}{A character string specifying the trait to be analyzed}

\item{subgroup}{A character string specifying the column containing the grouping variable}
}
\value{
A frequency table with the specified subgroup as the rownames, the scores of the specified trait as column names, and count as values
}
\description{
This function will be used to convert the raw data from the database to count data that can be passed into the snell function.
}
\details{
This function groups the data by the specified subgroup and trait, and counts the occurrences for each combination. It then reshapes the data into a frequency table.
}
\examples{
buildfreqtable(data = mydata, trait = "weight", subgroup = "SnellCategory")

}
